Consolidated pie chart related files/widgets
