+++ lib/screens/alias_tab/alias_detailed_screen.dart
@@ -3,10 +3,10 @@ import 'package:anonaddy/models/alias/alias_data_model.dart';
 import 'package:anonaddy/screens/alias_tab/alias_default_recipient.dart';
 import 'package:anonaddy/shared_components/alias_created_at_widget.dart';
 import 'package:anonaddy/shared_components/alias_detail_list_tile.dart';
-import 'package:anonaddy/shared_components/alias_pie_chart.dart';
 import 'package:anonaddy/shared_components/constants/material_constants.dart';
 import 'package:anonaddy/shared_components/constants/official_anonaddy_strings.dart';
 import 'package:anonaddy/shared_components/custom_loading_indicator.dart';
+import 'package:anonaddy/shared_components/pie_chart/alias_pie_chart.dart';
 import 'package:anonaddy/shared_components/recipient_list_tile.dart';
 import 'package:anonaddy/state_management/providers/class_providers.dart';
 import 'package:anonaddy/utilities/confirmation_dialog.dart';
