+++ lib/screens/alias_tab/alias_tab.dart
@@ -1,11 +1,10 @@
 import 'package:anonaddy/models/alias/alias_data_model.dart';
 import 'package:anonaddy/shared_components/alias_list_tile.dart';
-import 'package:anonaddy/shared_components/alias_tab_pie_chart.dart';
 import 'package:anonaddy/shared_components/constants/material_constants.dart';
 import 'package:anonaddy/shared_components/custom_page_route.dart';
 import 'package:anonaddy/shared_components/lottie_widget.dart';
+import 'package:anonaddy/shared_components/pie_chart/alias_tab_pie_chart.dart';
 import 'package:anonaddy/shared_components/shimmer_effects/alias_shimmer_loading.dart';
-import 'package:anonaddy/state_management/providers/class_providers.dart';
 import 'package:anonaddy/state_management/providers/global_providers.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -20,7 +19,6 @@ class AliasTab extends ConsumerWidget {
     /// preloads domainOptions for create new alias screen
     watch(domainOptionsProvider);
 
-    final aliasStateProvider = context.read(aliasStateManagerProvider);
     final size = MediaQuery.of(context).size;
 
     final List<AliasDataModel> availableAliasList = [];
