+++ lib/shared_components/pie_chart/alias_pie_chart.dart
@@ -1,5 +1,5 @@
 import 'package:anonaddy/models/alias/alias_data_model.dart';
-import 'package:anonaddy/shared_components/pie_chart_indicator.dart';
+import 'package:anonaddy/shared_components/pie_chart/pie_chart_indicator.dart';
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
